Improving the infrastructure for review environments:
    - the infrastructure provisioning has been split: shared infrastructure and PR specific
    - shared infrastructure: VPC, subnets, gateway, EC2 instance
    - PR specific: DB, Redis, ECS cluster
    - a Route53 subdomain is also created
