+++ .github/workflows/ci-review.yml
@@ -68,44 +68,44 @@ jobs:
           python-version: 3.7
           cache: "pip"
 
-      - name: Setup Env
-        run: |
-          echo "PYTHONPATH=/home/runner/work/web/web/app" >> $GITHUB_ENV
-          cp app/app/ci.env app/app/.env
-          pip install pip==20.0.2 setuptools wheel --upgrade
+      # - name: Setup Env
+      #   run: |
+      #     echo "PYTHONPATH=/home/runner/work/web/web/app" >> $GITHUB_ENV
+      #     cp app/app/ci.env app/app/.env
+      #     pip install pip==20.0.2 setuptools wheel --upgrade
 
-      - name: Fetch and Install GeoIP database files
-        run: |
-          sudo apt-get update && sudo apt-get install -y libmaxminddb-dev libsodium-dev libsecp256k1-dev
-          cp dist/*.gz ./
-          gunzip GeoLite2-City.mmdb.tar.gz && gunzip GeoLite2-Country.mmdb.tar.gz
-          tar -xvf GeoLite2-City.mmdb.tar && tar -xvf GeoLite2-Country.mmdb.tar
-          sudo mkdir -p /opt/GeoIP/
-          sudo mv GeoLite2-City_20200128/*.mmdb /opt/GeoIP/
-          sudo mv GeoLite2-Country_20200128/*.mmdb /opt/GeoIP/
+      # - name: Fetch and Install GeoIP database files
+      #   run: |
+      #     sudo apt-get update && sudo apt-get install -y libmaxminddb-dev libsodium-dev libsecp256k1-dev
+      #     cp dist/*.gz ./
+      #     gunzip GeoLite2-City.mmdb.tar.gz && gunzip GeoLite2-Country.mmdb.tar.gz
+      #     tar -xvf GeoLite2-City.mmdb.tar && tar -xvf GeoLite2-Country.mmdb.tar
+      #     sudo mkdir -p /opt/GeoIP/
+      #     sudo mv GeoLite2-City_20200128/*.mmdb /opt/GeoIP/
+      #     sudo mv GeoLite2-Country_20200128/*.mmdb /opt/GeoIP/
 
-      - name: Install libvips, Node, and Python dependencies
-        run: |
-          sudo apt-get install -y libvips libvips-dev
-          node --version
-          yarn install
-          pip install -r requirements/test.txt
-          yarn run eslint
-          yarn run stylelint
-          (cd app; python ./manage.py collectstatic --noinput --disable-collectfast)
+      # - name: Install libvips, Node, and Python dependencies
+      #   run: |
+      #     sudo apt-get install -y libvips libvips-dev
+      #     node --version
+      #     yarn install
+      #     pip install -r requirements/test.txt
+      #     yarn run eslint
+      #     yarn run stylelint
+      #     (cd app; python ./manage.py collectstatic --noinput --disable-collectfast)
 
-      - name: Run management commands
-        run: |
-          python app/manage.py migrate
-          python app/manage.py fetch_gas_prices
+      # - name: Run management commands
+      #   run: |
+      #     python app/manage.py migrate
+      #     python app/manage.py fetch_gas_prices
 
-      - name: Run Python and UI tests
-        run: |
-          pytest -p no:ethereum -p no:warnings
-          bin/ci/cypress-run
+      # - name: Run Python and UI tests
+      #   run: |
+      #     pytest -p no:ethereum -p no:warnings
+      #     bin/ci/cypress-run
 
-      - name: Generate Markdown documentation and static docs page
-        run: pydocmd build
+      # - name: Generate Markdown documentation and static docs page
+      #   run: pydocmd build
 
       - name: Deploy to Github Pages 🚀
         uses: peaceiris/actions-gh-pages@v3
@@ -187,12 +187,13 @@ jobs:
 
       # Run pulumi actions
       - uses: pulumi/actions@v3
-        id: pulumi
+        id: pulumi-env
+        name: Ensure the pulumi generic environment exist 
         with:
           command: up
-          stack-name: ${{ steps.compute.outputs.pulumi_stack }}
+          stack-name: review-env
           upsert: true
-          work-dir: infra/infra-review
+          work-dir: infra/review-env
         env:
           PULUMI_ACCESS_TOKEN: ${{ secrets.PULUMI_ACCESS_TOKEN }}
           PULUMI_CONFIG_PASSPHRASE: 
@@ -205,6 +206,39 @@ jobs:
           POC_PUBLIC_KEY_GE: ${{ secrets.POC_PUBLIC_KEY_GE }}
           POC_PUBLIC_KEY_GR: ${{ secrets.POC_PUBLIC_KEY_GR }}
           POC_DOCKER_GTC_WEB_IMAGE: ${{ needs.build-and-test.outputs.dockerTag }}
+
+      # Run pulumi actions
+      - uses: pulumi/actions@v3
+        id: pulumi
+        name: Create the environment specific for this PR
+        with:
+          command: up
+          stack-name: ${{ steps.compute.outputs.pulumi_stack }}
+          upsert: true
+          work-dir: infra/infra-pr
+        env:
+          PULUMI_ACCESS_TOKEN: ${{ secrets.PULUMI_ACCESS_TOKEN }}
+          PULUMI_CONFIG_PASSPHRASE: 
+          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+          AWS_REGION: ${{ secrets.AWS_REGION }}
+          POC_DB_NAME: ${{ secrets.POC_DB_NAME }}
+          POC_DB_PASSWORD: ${{ secrets.POC_DB_PASSWORD }}
+          POC_DB_USER: ${{ secrets.POC_DB_USER }}
+          POC_PUBLIC_KEY_GE: ${{ secrets.POC_PUBLIC_KEY_GE }}
+          POC_PUBLIC_KEY_GR: ${{ secrets.POC_PUBLIC_KEY_GR }}
+          POC_DOCKER_GTC_WEB_IMAGE: ${{ needs.build-and-test.outputs.dockerTag }}
+
+          # Pass in the outputs from the generic environment, we will deploy our ressources in the VPC
+          # and subnet that where created there
+          REVIEW_ENV_VPC_ID: ${{ steps.pulumi-env.outputs.vpcID }}
+          REVIEW_ENV_PRIVATE_SUBNET_1: ${{ steps.pulumi-env.outputs.vpcPrivateSubnetId1 }}
+          REVIEW_ENV_PRIVATE_SUBNET_2: ${{ steps.pulumi-env.outputs.vpcPrivateSubnetId2 }}
+          REVIEW_ENV_PUBLIC_SUBNET_1: ${{ steps.pulumi-env.outputs.vpcPublicSubnetId1 }}
+          REVIEW_ENV_PUBLIC_SUBNET_2: ${{ steps.pulumi-env.outputs.vpcPublicSubnetId2 }}
+
+          REVIEW_ENV_ROUTE53_ZONE_ID: ${{ secrets.REVIEW_ENV_ROUTE53_ZONE_ID }}
+          REVIEW_ENV_DOMAIN: review_123.gtc.iakobinyi.com
 
       - name: Start migration task
         run: |
