--- infra/infra-review/Pulumi.yaml
@@ -1,3 +1,0 @@
-name: infra-review
-runtime: nodejs
-description: Infrastructure for review env
