+++ infra/review-env/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/node_modules/
