+++ infra/review-env/Pulumi.yaml
@@ -1,0 +1,3 @@
+name: review-env
+runtime: nodejs
+description: Review environment shared ressources
