+++ infra/review-pr/Pulumi.yaml
@@ -1,0 +1,3 @@
+name: review
+runtime: nodejs
+description: Review environment for PR
