+++ infra/review-pr/index.ts
@@ -3,18 +3,28 @@ import * as aws from "@pulumi/aws";
 import * as awsx from "@pulumi/awsx";
 
 // The following vars ar not alloed to be undefined, hence the `${...}` magic
-let dbUsername = `${process.env["POC_DB_USER"]}`;
-let dbPassword = pulumi.secret(`${process.env["POC_DB_PASSWORD"]}`);
-let dbName = `${process.env["POC_DB_NAME"]}`;
+const dbUsername = `${process.env["POC_DB_USER"]}`;
+const dbPassword = pulumi.secret(`${process.env["POC_DB_PASSWORD"]}`);
+const dbName = `${process.env["POC_DB_NAME"]}`;
 
-let githubApiUser = `${process.env["POC_GITHUB_API_USER"]}`;
-let githubApiToken = pulumi.secret(`${process.env["POC_GITHUB_API_TOKEN"]}`);
-let githubClientId = `${process.env["POC_GITHUB_CLIENT_ID"]}`;
-let githubClientSecret = pulumi.secret(`${process.env["POC_GITHUB_CLIENT_SECRET"]}`);
-let githubAppName = `${process.env["POC_GITHUB_APP_NAME"]}`;
+const githubApiUser = `${process.env["POC_GITHUB_API_USER"]}`;
+const githubApiToken = pulumi.secret(`${process.env["POC_GITHUB_API_TOKEN"]}`);
+const githubClientId = `${process.env["POC_GITHUB_CLIENT_ID"]}`;
+const githubClientSecret = pulumi.secret(`${process.env["POC_GITHUB_CLIENT_SECRET"]}`);
+const githubAppName = `${process.env["POC_GITHUB_APP_NAME"]}`;
 
 
 export const dockerGtcWebImage = `${process.env["POC_DOCKER_GTC_WEB_IMAGE"]}`;
+
+
+const vpcID = `${process.env["REVIEW_ENV_VPC_ID"]}`;
+const privateSubnet1ID = `${process.env["REVIEW_ENV_PRIVATE_SUBNET_1"]}`;
+const privateSubnet2ID = `${process.env["REVIEW_ENV_PRIVATE_SUBNET_2"]}`;
+const publicSubnet1ID = `${process.env["REVIEW_ENV_PUBLIC_SUBNET_1"]}`;
+const publicSubnet2ID = `${process.env["REVIEW_ENV_PUBLIC_SUBNET_2"]}`;
+
+const route53ZoneID =  `${process.env["REVIEW_ENV_ROUTE53_ZONE_ID"]}`; 
+const domain =  `${process.env["REVIEW_ENV_DOMAIN"]}`; 
 
 
 //////////////////////////////////////////////////////////////
@@ -94,21 +104,14 @@ export const bucketWebURL = pulumi.interpolate`http://${staticAssetsBucket.websi
 // Set up VPC
 //////////////////////////////////////////////////////////////
 
-const vpc = new awsx.ec2.Vpc("gitcoin", {
-    subnets: [
-        { type: "public", },
-        { type: "private", mapPublicIpOnLaunch: true },
-    ],
+const vpc = awsx.ec2.Vpc.fromExistingIds("gitcoin", {
+    vpcId: vpcID,                                            
+    privateSubnetIds: [privateSubnet1ID, privateSubnet2ID],  
+    publicSubnetIds: [publicSubnet1ID, publicSubnet2ID]          
 });
 
-export const vpcID = vpc.id;
+export const vpc_id  = vpc.id;
 export const vpcPrivateSubnetIds = vpc.privateSubnetIds;
-export const vpcPublicSubnetIds = vpc.publicSubnetIds;
-
-
-export const vpcPublicSubnet1 = vpcPublicSubnetIds.then((subnets) => {
-    return subnets[0];
-});
 
 
 //////////////////////////////////////////////////////////////
@@ -188,9 +191,9 @@ export const redisPrimaryNode = redis.cacheNodes[0];
 export const redisConnectionUrl = pulumi.interpolate`rediscache://${redisPrimaryNode.address}:${redisPrimaryNode.port}/0?client_class=django_redis.client.DefaultClient`
 export const redisCacheOpsConnectionUrl = pulumi.interpolate`redis://${redisPrimaryNode.address}:${redisPrimaryNode.port}/0`
 
-//////////////////////////////////////////////////////////////
-// Set up ALB and ECS cluster
-//////////////////////////////////////////////////////////////
+// //////////////////////////////////////////////////////////////
+// // Set up ALB and ECS cluster
+// //////////////////////////////////////////////////////////////
 
 const cluster = new awsx.ecs.Cluster("gitcoin", { vpc });
 // export const clusterInstance = cluster;
@@ -551,90 +554,22 @@ export const frontendURL = pulumi.interpolate`http://${listener.endpoint.hostnam
 export const frontend = listener.endpoint
 
 
-//////////////////////////////////////////////////////////////
-// Set up EC2 instance 
-//      - it is intended to be used for troubleshooting
-//////////////////////////////////////////////////////////////
-
-// Create a new security group that permits SSH and web access.
-const secgrp = new aws.ec2.SecurityGroup("secgrp", {
-    description: "gitcoin",
-    vpcId: vpc.id,
-    ingress: [
-        { protocol: "tcp", fromPort: 22, toPort: 22, cidrBlocks: ["0.0.0.0/0"] },
-        { protocol: "tcp", fromPort: 80, toPort: 80, cidrBlocks: ["0.0.0.0/0"] },
-    ],
-    egress: [{
-        protocol: "-1",
-        fromPort: 0,
-        toPort: 0,
-        cidrBlocks: ["0.0.0.0/0"],
-    }],
+const www = new aws.route53.Record("www", {
+    zoneId: route53ZoneID,    // TODO: hardcoded
+    name: domain,     // TODO: hardcoded
+    type: "CNAME",
+    ttl: 300,
+    records: [listener.endpoint.hostname],
 });
 
-export const securityGroupsForEc2 = secgrp.id;
-
-const ubuntu = aws.ec2.getAmi({
-    mostRecent: true,
-    filters: [
-        {
-            name: "name",
-            values: ["ubuntu/images/hvm-ssd/ubuntu-focal-20.04-amd64-server-*"],
-        },
-        {
-            name: "virtualization-type",
-            values: ["hvm"],
-        },
-    ],
-    owners: ["099720109477"],
-});
-
-// Script to install docker in ec2 instance
-const ec2InitScript = `#!/bin/bash
-
-# Installing docker in ubuntu
-# Instructions taken from here: https://docs.docker.com/engine/install/ubuntu/
-
-mkdir /var/log/gitcoin
-echo $(date) "Starting installation of docker" >> /var/log/gitcoin/init.log
-apt-get remove docker docker-engine docker.io containerd runc
-
-apt-get update
-
-apt-get install -y \
-    ca-certificates \
-    curl \
-    gnupg \
-    lsb-release
-    
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
-
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu \
-  $(lsb_release -cs) stable" | tee /etc/apt/sources.list.d/docker.list > /dev/null
-  
-apt-get update
-apt-get install -y docker-ce docker-ce-cli containerd.io
-mkdir /var/log/gitcoin
-echo $(date) "Finished installation of docker" >> /var/log/gitcoin/init.log
-
+// Build and export the command to use the docker image of this specific deployment from an EC2 instance within the subnet
+export const dockerConnectCmd = pulumi.interpolate`docker run -it \
+-e DATABASE_URL=${rdsConnectionUrl} \
+-e CACHEOPS_REDIS=${redisCacheOpsConnectionUrl} \
+-e REDIS_URL=${redisConnectionUrl} \
+-e STATIC_HOST=${bucketWebURL} \
+-e STATIC_URL=static/ \
+-e STATICFILES_STORAGE=django.contrib.staticfiles.storage.StaticFilesStorage \
+-e ENV=test \
+${dockerGtcWebImage} bash \
 `
-
-const web = new aws.ec2.Instance("Web", {
-    ami: ubuntu.then(ubuntu => ubuntu.id),
-    associatePublicIpAddress: true,
-    instanceType: "t3.micro",
-    subnetId: vpcPublicSubnet1.then(),
-
-    vpcSecurityGroupIds: [secgrp.id],
-    rootBlockDevice: {
-        volumeSize: 50
-    },
-    tags: {
-        Name: "Troubleshooting instance",
-    },
-    userData: ec2InitScript,
-});
-
-export const ec2PublicIp = web.publicIp;
-
